+++ scrapeomatic/utils/async_utils.py
@@ -1,6 +1,7 @@
 import logging
 
 from playwright.sync_api import Route
+from selenium import webdriver
 
 from scrapeomatic.utils.constants import PLAYWRIGHT_BLOCK_RESOURCE_TYPES, PLAYWRIGHT_BLOCK_RESOURCE_NAMES
 
@@ -23,3 +24,20 @@ class AsyncUtils:
             logging.debug(f"blocking background resource {route.request} blocked name {route.request.url}")
             return route.abort()
         return route.continue_()
+
+    @staticmethod
+    def get_cookie_dict(driver: webdriver) -> dict:
+        """
+        Gets the cookies from Selenium webdriver and reformats them into a dictionary.
+        Args:
+            driver: The input webdriver object
+
+        Returns: A dictionary of cookies.
+
+        """
+        all_cookies = driver.get_cookies()
+        cookies_dict = {}
+        for cookie in all_cookies:
+            cookies_dict[cookie['name']] = cookie['value']
+
+        return cookies_dict
